Update streamlit_app.py
After commiting the new line to the file, begin interacting with all three areas of the page. Add and take away fruits from the [Build Your Own Fruit Smoothie] picker. Enter values into the Fruityvice box and do the same in your new box at the bottom of the page.